rename RxRangeFacet to RxRangeFilter
